# frontend/contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "testhost | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/reports/{reportKey}/embed-token": {
      "post": {
        "tags": [
          "Reports"
        ],
        "summary": "Issue a short-lived Superset guest token for a report",
        "operationId": "CreateReportEmbedToken",
        "parameters": [
          {
            "name": "reportKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportEmbedTokenResult"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ReportEmbedTokenResult": {
        "required": [
          "reportKey",
          "embedUrl",
          "token",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "reportKey": {
            "type": "string"
          },
          "embedUrl": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}